Cleaned up old settings code #10
